refact: split apart ingest file into a multi-file module
