+++ MANIFEST.in
@@ -1,4 +1,4 @@
-include chord_metadata_service/chord/workflows/*.wdl
+include chord_metadata_service/chord/workflows/wdls/*.wdl
 include chord_metadata_service/chord/tests/*.json
 include chord_metadata_service/dats/*
 include chord_metadata_service/mcode/tests/*.json
