+++ chord_metadata_service/chord/ingest/exceptions.py
@@ -1,0 +1,7 @@
+__all__ = [
+    "IngestError",
+]
+
+
+class IngestError(Exception):
+    pass
