+++ chord_metadata_service/chord/ingest/logger.py
@@ -1,0 +1,5 @@
+import logging
+
+__all__ = ["logger"]
+
+logger = logging.getLogger(__name__)
