+++ chord_metadata_service/chord/ingest/mcode.py
@@ -1,0 +1,26 @@
+import json
+from chord_metadata_service.mcode.mcode_ingest import ingest_mcodepacket
+from chord_metadata_service.mcode.parse_fhir_mcode import parse_bundle
+from .logger import logger
+from .utils import get_output_or_raise, workflow_file_output_to_path
+
+
+def ingest_mcode_fhir_workflow(workflow_outputs, table_id):
+    with workflow_file_output_to_path(get_output_or_raise(workflow_outputs, "json_document")) as json_doc_path:
+        logger.info(f"Attempting ingestion of MCODE FIHR from path: {json_doc_path}")
+        with open(json_doc_path, "r") as jf:
+            json_data = json.load(jf)
+            mcodepacket = parse_bundle(json_data)
+            ingest_mcodepacket(mcodepacket, table_id)
+
+
+def ingest_mcode_workflow(workflow_outputs, table_id):
+    with workflow_file_output_to_path(get_output_or_raise(workflow_outputs, "json_document")) as json_doc_path:
+        logger.info(f"Attempting ingestion of MCODE from path: {json_doc_path}")
+        with open(json_doc_path, "r") as jf:
+            json_data = json.load(jf)
+            if isinstance(json_data, list):
+                for mcodepacket in json_data:
+                    ingest_mcodepacket(mcodepacket, table_id)
+            else:
+                ingest_mcodepacket(json_data, table_id)
