+++ chord_metadata_service/chord/tests/example_ingest.py
@@ -2,10 +2,18 @@ import json
 import os
 
 
-__all__ = ["EXAMPLE_INGEST_PHENOPACKET", "EXAMPLE_INGEST_OUTPUTS",
-           "EXAMPLE_INGEST_EXPERIMENT", "EXAMPLE_INGEST_OUTPUTS_EXPERIMENT",
-           "EXAMPLE_INGEST_INVALID_EXPERIMENT", "EXAMPLE_INGEST_INVALID_PHENOPACKET",
-           "EXAMPLE_INGEST_MULTIPLE_PHENOPACKETS", "EXAMPLE_INGEST_MULTIPLE_OUTPUTS"]
+__all__ = [
+    "EXAMPLE_INGEST_EXPERIMENT",
+    "EXAMPLE_INGEST_EXPERIMENT_RESULT",
+    "EXAMPLE_INGEST_PHENOPACKET",
+    "EXAMPLE_INGEST_OUTPUTS",
+    "EXAMPLE_INGEST_OUTPUTS_EXPERIMENT",
+    "EXAMPLE_INGEST_OUTPUTS_EXPERIMENT_RESULT",
+    "EXAMPLE_INGEST_INVALID_EXPERIMENT",
+    "EXAMPLE_INGEST_INVALID_PHENOPACKET",
+    "EXAMPLE_INGEST_MULTIPLE_PHENOPACKETS",
+    "EXAMPLE_INGEST_MULTIPLE_OUTPUTS"
+]
 
 with open(os.path.join(os.path.dirname(__file__), "example_phenopacket.json"), "r") as pf:
     EXAMPLE_INGEST_PHENOPACKET = json.load(pf)
