+++ chord_metadata_service/chord/workflows/constants.py
@@ -1,0 +1,1 @@
+FROM_CONFIG = "FROM_CONFIG"
