+++ chord_metadata_service/mcode/tests/test_ingest.py
@@ -7,11 +7,8 @@ from chord_metadata_service.chord.data_types import DATA_TYPE_MCODEPACKET
 from chord_metadata_service.chord.models import Project, Dataset, TableOwnership, Table
 from chord_metadata_service.patients.models import Individual
 from chord_metadata_service.phenopackets.models import Gene
-# noinspection PyProtectedMember
-from chord_metadata_service.chord.ingest import (
-    WORKFLOW_INGEST_FUNCTION_MAP,
-    WORKFLOW_MCODE_JSON
-)
+from chord_metadata_service.chord.ingest import WORKFLOW_INGEST_FUNCTION_MAP
+from chord_metadata_service.chord.workflows.metadata import WORKFLOW_MCODE_JSON
 from chord_metadata_service.chord.tests.constants import VALID_DATA_USE_1
 
 from ..models import (
