+++ chord_metadata_service/restapi/tests/test_fhir_ingest.py
@@ -6,14 +6,12 @@ from rest_framework.test import APITestCase
 
 from chord_metadata_service.chord.data_types import DATA_TYPE_PHENOPACKET
 from chord_metadata_service.chord.models import Project, Dataset, TableOwnership, Table
+from chord_metadata_service.chord.ingest import WORKFLOW_INGEST_FUNCTION_MAP
+from chord_metadata_service.chord.tests.constants import VALID_DATA_USE_1
+from chord_metadata_service.chord.workflows.metadata import WORKFLOW_FHIR_JSON
 from chord_metadata_service.patients.models import Individual
 from chord_metadata_service.phenopackets.models import PhenotypicFeature, Biosample, Disease, Procedure, Phenopacket
-from chord_metadata_service.chord.tests.constants import VALID_DATA_USE_1
 from chord_metadata_service.restapi.fhir_ingest import ingest_patients, ingest_observations
-from chord_metadata_service.chord.ingest import (
-    WORKFLOW_INGEST_FUNCTION_MAP,
-    WORKFLOW_FHIR_JSON
-)
 from .constants import INVALID_FHIR_BUNDLE_1, INVALID_SUBJECT_NOT_PRESENT
